Made major refactoring changes to the codebase.
-> Moved the Database Connection String to the Web.Config file and out
of the application code.

-> Moved all sql statement out of the code and into stored procedures in
the database.

-> Implemented the use of SqlParameter objects to pass parameters into
the stored procedures - in place of simple string concatenation which is
a gaping security hole.

-> Implemented a DataAccess class so that all Database related code can
be stored and called from a single location. This class includes
generous use of Using{} statements to ensure proper disposal of all
managed resources used when making database calls.

-> Overloaded the constructor on the Product class to allow the class to
be instantiated without any database calls within the class

-> Overloaded the constructor on the ProductOption class to allow the class to
be instantiated without any database calls within the class

-> Modified the Products class LoadProducts method so that a single
database query is needed to get all data and instantiate all Product
objects - instead of 1 query to get the ids and then multiple other
queries to instantiate the objects.

-> Modified the ProductOptions class LoadProductOptions method so that a single
database query is needed to get all data and instantiate all Product
objects - instead of 1 query to get the ids and then multiple other
queries to instantiate the objects.

-> Change the any variables declared as "var" to their proper type.

-> Modified all api methods in the ProductController so that they return
IHttpActionResult.

-> All api methods in the ProductController return Ok(<T>) when
execution is successful.

-> Impletmented error handling in all api methods to throw an
InternalServerError on any unhandled exceptions.

-> Added XML Doc comments on all methods in the ProductController.

-> Created a Helper namespace and folder within the project.

-> Moved the Helper.cs class into the new namespace and folder.

-> Implemented a bool variable ProductFound in the Product class for
indicating whether a given ProductId was found in the databasse.

-> Implemented a bool variable ProductOptionFound in the ProductOption
class for indicating whether a given ProductOptionId was found in the
database.